requirements.txt
fastapi==0.78.0
jupyter==1.0.0
Pillow==9.2.0
timm==0.6.5
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
uvicorn==0.17.6
uvloop==0.16.0
transformers==4.20.1